Use test-level tags in CI
Assigns a test level to each integration test and ensures that the integration tests
fail if a tag is missing for a test.
Also, enables the tags in the CI
