+++ .github/workflows/integration-informing.yaml
@@ -94,7 +94,7 @@ jobs:
           TEST_FLAVOR: ${{ matrix.patch }}
           TEST_INSPECTION_REPORTS_DIR: ${{ github.workspace }}/inspection-reports
         run: |
-          cd tests/integration && sg lxd -c 'tox -e integration'
+          cd tests/integration && sg lxd -c 'tox -e integration -- --tags pull_request'
       - name: Prepare inspection reports
         if: failure()
         run: |
