+++ tests/integration/tests/test_bootstrap.py
@@ -4,11 +4,12 @@
 from typing import List
 
 import pytest
-from test_util import harness
+from test_util import harness, tags
 
 
 @pytest.mark.node_count(1)
 @pytest.mark.disable_k8s_bootstrapping()
+@pytest.mark.tags(tags.NIGHTLY)
 def test_microk8s_installed(instances: List[harness.Instance]):
     instance = instances[0]
     instance.exec("snap install microk8s --classic".split())
