+++ tests/integration/tests/test_etcd.py
@@ -7,7 +7,7 @@ from typing import List
 
 import pytest
 import yaml
-from test_util import harness, util
+from test_util import harness, util, tags
 from test_util.etcd import EtcdCluster
 
 LOG = logging.getLogger(__name__)
@@ -16,6 +16,7 @@ LOG = logging.getLogger(__name__)
 @pytest.mark.node_count(1)
 @pytest.mark.etcd_count(1)
 @pytest.mark.disable_k8s_bootstrapping()
+@pytest.mark.tags(tags.NIGHTLY)
 def test_etcd(instances: List[harness.Instance], etcd_cluster: EtcdCluster):
     k8s_instance = instances[0]
