+++ tests/integration/tests/test_ingress.py
@@ -9,7 +9,7 @@ from pathlib import Path
 from typing import List
 
 import pytest
-from test_util import config, harness, util
+from test_util import config, harness, util, tags
 from test_util.config import MANIFESTS_DIR
 
 LOG = logging.getLogger(__name__)
@@ -74,6 +74,7 @@ def get_external_service_ip(instance: harness.Instance, service_namespace) -> st
 
 
 @pytest.mark.bootstrap_config((config.MANIFESTS_DIR / "bootstrap-all.yaml").read_text())
+@pytest.mark.tags(tags.PULL_REQUEST)
 def test_ingress(instances: List[harness.Instance]):
     instance = instances[0]
     instance_default_ip = util.get_default_ip(instance)
