+++ tests/integration/tests/test_networking.py
@@ -6,7 +6,7 @@ from ipaddress import IPv4Address, IPv6Address, ip_address
 from typing import List
 
 import pytest
-from test_util import config, harness, util
+from test_util import config, harness, util, tags
 
 LOG = logging.getLogger(__name__)
 
@@ -16,6 +16,7 @@ LOG = logging.getLogger(__name__)
     (config.MANIFESTS_DIR / "bootstrap-dualstack.yaml").read_text()
 )
 @pytest.mark.dualstack()
+@pytest.mark.tags(tags.NIGHTLY)
 def test_dualstack(instances: List[harness.Instance]):
     main = instances[0]
     dualstack_config = (config.MANIFESTS_DIR / "nginx-dualstack.yaml").read_text()
@@ -61,6 +62,7 @@ def test_dualstack(instances: List[harness.Instance]):
 @pytest.mark.node_count(3)
 @pytest.mark.disable_k8s_bootstrapping()
 @pytest.mark.network_type("dualstack")
+@pytest.mark.tags(tags.NIGHTLY)
 def test_ipv6_only_on_dualstack_infra(instances: List[harness.Instance]):
     main = instances[0]
     joining_cp = instances[1]
